refactor: move codebase to app/ package (no logic change)
